+++ pkg/apis/configuration/v1/kongconsumer_types.go
@@ -61,6 +61,16 @@ type KongConsumerList struct {
 
 // KongConsumerStatus represents the current status of the KongConsumer resource.
 type KongConsumerStatus struct {
+	// Conditions describe the current conditions of the KongConsumer.
+	//
+	// Known condition types are:
+	//
+	// * "Programmed"
+	//
+	// +listType=map
+	// +listMapKey=type
+	// +kubebuilder:validation:MaxItems=8
+	// +kubebuilder:default={{type: "Programmed", status: "Unknown", reason:"Pending", message:"Waiting for controller", lastTransitionTime: "1970-01-01T00:00:00Z"}}
 	Conditions []metav1.Condition `json:"conditions,omitempty"`
 }